+++ test/integration/playwright.config.ts
@@ -1,12 +1,13 @@
 import { PlaywrightTestConfig, devices } from '@playwright/test';
 
-const config: PlaywrightTestConfig = {
+const config: PlaywrightTestConfig<{ toolpadDev: boolean }> = {
   forbidOnly: !!process.env.CI,
   retries: process.env.CI ? 2 : 0,
   workers: 1,
   use: {
     trace: 'on-first-retry',
     baseURL: process.env.PLAYWRIGHT_TEST_BASE_URL || 'http://localhost:3000/',
+    toolpadDev: !!process.env.TOOLPAD_DEV,
   },
   globalSetup: '../playwright/global-setup',
   globalTeardown: '../playwright/global-teardown',
